requirements.txt
openai
python-dotenv
pandas
requests
python-binance
